outsidetemp.old/requirements.txt
Flask==1.0
flask-restful==0.3.6
python-dateutil==2.7.2